Split up the class per feedback
