+++ WheelWizard/Features/WiiManagement/Domain/Mii/MiiEnums.cs
@@ -1,0 +1,20 @@
+﻿namespace WheelWizard.WiiManagement.Domain.Mii;
+
+public enum MiiFavoriteColor : uint { Red, Orange, Yellow, Green, Blue, LightBlue, Pink, Purple, Brown, White, Black, Gray }
+public enum MiiFaceShape { RoundPointChin, Circle, Oval, BlobFatChin, RightAnglePointChin, Bread, Octagon, Square}
+public enum MiiSkinColor { Light, LightTan, Tan, Pink, DarkBrown, Brown}
+public enum MiiFacialFeature {None, Cheeks, CheekAndEyes, Freckles, BaggyEyes, Chad, Tired, Chin, EyeShadow, Beard, MouthCorners, Old}
+public enum HairColor {Black, Brown, Red, LightRed, Grey, LightBrown, Blonde, White}
+public enum EyebrowColor {Black, Brown, Red, LightRed, Grey, LightBrown, Blonde, White}
+public enum EyeColor : uint {Black, Grey, Red, Gold, Blue, Green}
+public enum NoseType  {Default, SemiCircle, Dots, VShape, FullNose, Triangle, FlatC, UpsideDownC, Squidward, ArrowDown, Flat, Tunnel}
+public enum LipColor {Skin, Red, Pink}
+public enum GlassesColor  {Dark, DarkGold, Red, Blue, Gold, White}
+public enum GlassesType  {None, Square, Rectangle, Circle, Oval, Misses, SadSunGlasses, SunGlasses, CoolSunGlasses}
+public enum MustacheColor {Black, Brown, Red, LightRed, Grey, LightBrown, Blonde, White}
+public enum MustacheType {None, Fat, Thin, Goatee}
+public enum BeardType {None, Thin, Wide, Widest}
+
+
+
+
