+++ WheelWizard/Models/RRInfo/RrPlayer.cs
@@ -1,4 +1,5 @@
 using WheelWizard.WheelWizardData.Domain;
+using WheelWizard.WiiManagement.Domain.Mii;
 
 
 namespace WheelWizard.Models.RRInfo;
