+++ WheelWizard/Models/RRInfo/RrRoom.cs
@@ -1,5 +1,6 @@
 ﻿using WheelWizard.Helpers;
 using WheelWizard.Models.MiiImages;
+using WheelWizard.WiiManagement.Domain.Mii;
 
 namespace WheelWizard.Models.RRInfo;
