+++ WheelWizard/Views/Components/WhWzLibrary/DetailedProfileBox.axaml.cs
@@ -7,6 +7,7 @@ using System.ComponentModel;
 using WheelWizard.Models.MiiImages;
 using WheelWizard.Services.Settings;
 using WheelWizard.Views.Components.MiiImages;
+using WheelWizard.WiiManagement.Domain.Mii;
 
 namespace WheelWizard.Views.Components;
