+++ WheelWizard/Views/Components/WhWzLibrary/FriendsListItem.axaml.cs
@@ -4,6 +4,7 @@ using Avalonia.Controls.Primitives;
 using Avalonia.Interactivity;
 using WheelWizard.Models.MiiImages;
 using WheelWizard.WheelWizardData;
+using WheelWizard.WiiManagement.Domain.Mii;
 
 namespace WheelWizard.Views.Components;
