+++ WheelWizard/Views/Components/WhWzLibrary/PlayerListItem.axaml.cs
@@ -3,6 +3,7 @@ using Avalonia.Controls;
 using Avalonia.Controls.Primitives;
 using WheelWizard.Models.MiiImages;
 using WheelWizard.WheelWizardData;
+using WheelWizard.WiiManagement.Domain.Mii;
 
 
 namespace WheelWizard.Views.Components;
